Add fix-path option.
On OS X electron frequently does not seem to pick up the PATH setting
correctly when launching Atom from a GUI, for example the Dock. This
means that pep8 can not be found. One workaround is to specify the
absolute path to pep8, but this can be done automatically. This
provides a much better experience for users on OS X.

- Adds `enableFixPath` option.
- Defaults `enableFixPath` to true for 'darwin` (OS X) platform.
- Provides user documentation for option, indicating its use in OS X.
- Adds optional `fix-path` dependency (avoid issues with non-OS X OS's).